backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
motor==3.6.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.18
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.37.0
google-auth-oauthlib==1.2.1
aiofiles==24.1.0
aiokafka==0.11.0
redis[hiredis]==5.2.1
# File extraction
PyPDF2==3.0.1
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5
tabulate==0.9.0
# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.27.0
mongomock-motor==0.0.31
fakeredis==2.21.1
httpx
pytest
pytest-asyncio